added timeout functionality for IAC supplementary info
