+++ model/src/main/java/uk/gov/hmcts/payment/api/dto/MissingSupplementaryDetailsDto.java
@@ -1,6 +1,5 @@
-package uk.gov.hmcts.payment.api.contract;
+package uk.gov.hmcts.payment.api.dto;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -8,25 +7,20 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uk.gov.hmcts.payment.api.model.PaymentAllocationStatus;
 
 import javax.validation.constraints.NotNull;
-import java.util.Date;
 import java.util.List;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 @JsonInclude(NON_NULL)
-@Builder(builderMethodName = "supplementaryInfoDtoWith")
+@Builder(builderMethodName = "MissingSupplementaryDetailsDtoWith")
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class SupplementaryInfoDto {
+public class MissingSupplementaryDetailsDto {
 
     @NotNull
-    private String ccdCaseNumber;
-
-    @NotNull
-    private SupplementaryDetailsDto supplementaryDetails;
+    private List<String> ccdCaseNumbers;
 }
