+++ model/src/main/java/uk/gov/hmcts/payment/api/dto/SupplementaryInfoDto.java
@@ -1,6 +1,5 @@
-package uk.gov.hmcts.payment.api.contract;
+package uk.gov.hmcts.payment.api.dto;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -8,22 +7,20 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uk.gov.hmcts.payment.api.model.PaymentAllocationStatus;
-
 import javax.validation.constraints.NotNull;
-import java.util.Date;
-import java.util.List;
-
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 @JsonInclude(NON_NULL)
-@Builder(builderMethodName = "supplementaryDetailsDtoWith")
+@Builder(builderMethodName = "supplementaryInfoDtoWith")
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class SupplementaryDetailsDto {
+public class SupplementaryInfoDto {
 
     @NotNull
-    private String surname;
+    private String ccdCaseNumber;
+
+    @NotNull
+    private SupplementaryDetailsDto supplementaryDetails;
 }
